==> 05_script_close_graphic.R <==

#################

cat(r"( ____        _               _                           _ _ 
|  _ \  __ _| |_ __ _    ___| | ___  __ _ _ __   ___  __| | |
| | | |/ _` | __/ _` |  / __| |/ _ \/ _` | '_ \ / _ \/ _` | |
| |_| | (_| | || (_| | | (__| |  __/ (_| | | | |  __/ (_| |_|
|____/ \__,_|\__\__,_|  \___|_|\___|\__,_|_| |_|\___|\__,_(_))")


cat("\n\n\n\n\n\n\n\n\n\n")


cat(r"() ____            _       _                               _      _       _ 
/ ___|  ___ _ __(_)_ __ | |_    ___ ___  _ __ ___  _ __ | | ___| |_ ___| |
\___ \ / __| '__| | '_ \| __|  / __/ _ \| '_ ` _ \| '_ \| |/ _ \ __/ _ \ |
 ___) | (__| |  | | |_) | |_  | (_| (_) | | | | | | |_) | |  __/ ||  __/_|
|____/ \___|_|  |_| .__/ \__|  \___\___/|_| |_| |_| .__/|_|\___|\__\___(_)
                  |_|                             |_|                     )")
Sys.sleep(2)

cat(r"(%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%@@%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
%%%%%%%%%%%@@@%%@@@@%%%@@@@%%@@@@@@@%%%@@@@%%%@@@@@@@@%@@@@@@@@%@@@@%@@@@@@@@%@@@%@@@@@@@%@@@%%%%%%%
%%%%%%%%%%%@@@%%@@@@@%@@@@@%%@@%%%@@%%%@@@@@%%@@@%%@@@%%%@@@%%%%%%%%%%@%%%%@@%@%%%%%%%%%%%%%%%%%%%%%
%%%%%%%%%%%@@@%%@@@@@%@@@@@%%@@@@@@@%%@@@%@@%%@@@%%%%%%%%@@@%%%%@@@@@@@@@@@@@@@@@@%@@@@@@@@@@@@@%%%%
%%%%%%%%%%%@@@%%@@%@@@@%@@@%%@@%%%%%%%@@@@@@@%@@@%%%@%%%%@@@%%%%@%@%@@@@@%@@@@@@@@%@@@@@%@@@%@%@%%%%
%%%%%%%%%%%@@@%%@@%%@@@%@@@%%@@%%%%%%@@@@@@@@%@@@%%@@@%%%@@@%%%%@%%%%%%%%%@%%%%%%@%%%%%%%%%%%%%%%%%%
%%%%%%%%%%%@@@%%@@%%@@@%@@@%%@@%%%%%%@@%%%%@@%%@@@@@@%%%%@@@%%%%@@@@@@@@@@@@@@@%%@@@@@@@@%%%%%%%%%%%
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%)")
Sys.sleep(0.2)
cat(r"(####################################################################################################
####################################################################################################
####################################################################################################
####################################################################################################
####################################################################################################
####################################################################################################
####################################################################################################
####################################################################################################
####################################################################################################
####################################################################################################
#################################################%%#################################################
###########%@@##@@@@###@@@%##@@@@@@@###@@@@###%@@@@@@%#@@@@@@@%#%%%%#%%%%%%%%#%%%#%%%%%%%%@%%#######
###########%@@##@@@@%#%@@@%##@@###@@###@@@@%##%@%##%@%###%@@##########%####%@#%#####################
###########%@@##@@%@@#@@%@%##@@@@@@%##@@%#@@##%@%########%@@####@%%%%%%%%%%%%%%%%%#%%%%%%%%%%%%%####
###########%@@##@@#@@@@#%@%##@@#######@@%%@@%#%@%###%####%@@####%#%#%%%%%#%%%%%%@%#%%%%%#%%%#%#%####
###########%@@##@@##@@@#%@%##@@######@@@@@@@@#%@@##@@%###%@@####@#########%######%##################
###########%@@##@@##@@%#%@%##@@######@@####@@##%@@@@%####%@@####@%@%@%@%@%@%%%%##%%%%%%%%###########
####################################################################################################
####################################################################################################
####################################################################################################
####################################################################################################
####################################################################################################
####################################################################################################
####################################################################################################
####################################################################################################
####################################################################################################
####################################################################################################
####################################################################################################)")
Sys.sleep(0.2)
cat(r"(****************************************************************************************************
****************************************************************************************************
****************************************************************************************************
****************************************************************************************************
****************************************************************************************************
****************************************************************************************************
****************************************************************************************************
****************************************************************************************************
****************************************************************************************************
****************************************************************************************************
***********###**####***####**#####*****####*****#%%#***########*###*****#***************************
***********%@@**@@@@**#@@@%*#@@%@@@%***@@@@#**#@@@@@@##@@@@@@@%*#%%%#%%%%%%%%#%%%#%%##%%%#@%%*******
***********%@@**@@%@#*%@%@%*#@@#*#@@#*#@@@@%**%@%**#%%***%@@****######%####%%###**#*#*#*#*#*#*******
***********%@@**@@#@@#@@#@%*#@@@@@@%**%@%#@@**%@%********%@@****@#%%%#%%%#%%##%%%#*%%%%%#%#%#%##****
***********%@@**@@#@@@@##@%*#@@##****#@@%%@@%*%@%**###***%@@****%#%#%%%%%##%%###%%*%%%%###%%#%#%****
***********%@@**@@##@@@*#@%*#@@#*****%@@%@@@@*%@@##@@%***%@@****%#########%#####*%#######***********
***********%@@**@@#*%@%*#@%*#@@#****#@@#**#@@#*%@@@@%****%@@****@%%%%%%%%%%%%%%#*%%%%%%%%***********
******************************************************************************#*********************
****************************************************************************************************
****************************************************************************************************
****************************************************************************************************
****************************************************************************************************
****************************************************************************************************
****************************************************************************************************
****************************************************************************************************
****************************************************************************************************
****************************************************************************************************
****************************************************************************************************)")
Sys.sleep(0.2)
cat(r"(****************************************************************************************************
****************************************************************************************************
****************************************************************************************************
****************************************************************************************************
****************************************************************************************************
****************************************************************************************************
****************************************************************************************************
****************************************************************************************************
****************************************************************************************************
****************************************************************************************************
***********###**####***####**#####*****####*****#%%#***########*###*****#***************************
***********%@@**@@@@**#@@@%*#@@%@@@%***@@@@#**#@@@@@@##@@@@@@@%*#%%%#%%%%%%%%#%%%#%%##%%%#@%%*******
***********%@@**@@%@#*%@%@%*#@@#*#@@#*#@@@@%**%@%**#%%***%@@****######%####%%###**#*#*#*#*#*#*******
***********%@@**@@#@@#@@#@%*#@@@@@@%**%@%#@@**%@%********%@@****@#%%%#%%%#%%##%%%#*%%%%%#%#%#%##****
***********%@@**@@#@@@@##@%*#@@##****#@@%%@@%*%@%**###***%@@****%#%#%%%%%##%%###%%*%%%%###%%#%#%****
***********%@@**@@##@@@*#@%*#@@#*****%@@%@@@@*%@@##@@%***%@@****%#########%#####*%#######***********
***********%@@**@@#*%@%*#@%*#@@#****#@@#**#@@#*%@@@@%****%@@****@%%%%%%%%%%%%%%#*%%%%%%%%***********
******************************************************************************#*********************
****************************************************************************************************
****************************************************************************************************
****************************************************************************************************
****************************************************************************************************
****************************************************************************************************
****************************************************************************************************
****************************************************************************************************
****************************************************************************************************
****************************************************************************************************
****************************************************************************************************)")
Sys.sleep(0.2)
cat(r"(...................................................................................................:
...................................................................................................:
...................................................................................................:
...................................................................................................:
...................................................................................................:
...................................................................................................:
...................................................................................................:
...................................................................................................:
...................................................................................................:
...................................................................................................:
...........:-:..---:...---:.:----:.....:--:....:-++-:..-------:.:::.....:.............::...........:
...........+@%..%@@#..-@@@+.-@@##@@#..:%@@%:..=@@##@@=:%%@@@%%=.=*+*-*+*+*+*+:+*+-+*==+++=%+*:.....:
...........+@%..%@#@=.*@#@*.-@%:.-@@:.-@%#@+..*@*..=#+...+@%....::::::+::::**:=-:::::::::.:::......:
...........+@%..%@=@#-@#=@*.-@@%%@@+..*@+-@%..*@*........+@%....#=+**=+=+=+===+++=.=++=*=+=+=+==...:
...........+@%..%@:#@%@-=@*.-@@-:....-@@++@@=.*@*..-=-...+@%....+-=-=+++=-=+====#+:*+++=-=+=-+-+...:
...........+@%..%@:-@@%:=@*.-@%:.....#@%###@%:+@%::#@*...+@#....#:-::::::-*-:::::*:::::::..........:
...........+@%..%@:.#@=.=@*.-@%:....:@@-..:%@-.+@@@@*....+@#....#=*+#*#+#+#+*++-.*++*+*+*..........:
....................................................................:.........::...................:
...................................................................................................:
...................................................................................................:
...................................................................................................:
...................................................................................................:
...................................................................................................:
...................................................................................................:
...................................................................................................:
...................................................................................................:
...................................................................................................:
...................................................................................................:)")
Sys.sleep(0.2)
cat(r"(                                                                                                    
                                                                                                    
                                                                                                    
                                                                                                    
                                                                                                    
                                                                                                    
                                                                                                    
                                                                                                    
                                                                                                    
                                                                                                    
                                                 ..                                                 
           :@*  #@@*   %@@-  @@++@@+   *@@%    @@*+@@  #%%@@%%.  --= =-=:=.=: .=: :- .:.: *-=       
           :@*  #@+@  -@+@-  @#   @@   @#*@.  =@=  .+:   :@*          :    -= .                     
           :@*  #@ @* @*.@-  @@##@@:  =@: @#  =@=        :@*    *.--= :.: :.  :.:  ....- . : :      
           :@*  #@ +@%@ .@-  @%       @@-:%@. =@=   .    :@*    : . ..:.. .:.   +: =...   :. : :    
           :@*  #@  @@# .@-  @%      +@#++*@* :@*  *@-   :@*    +         -      -                  
           :@#  #@  +@. .@-  @%      %%    #@  -@@@@-    :@*    *.=:+-+.+:+:=::  -::---.-           
                                                                                                    
)")

Sys.sleep(0.5)


cat(r"(                                                        =--_
                                         .-""""""-.     |* _)
                                        /          \   /  /
                                       /            \_/  /
           _                          /|                /
       _-'"/\                        / |    ____    _.-"            _
    _-'   (  '-_            _       (   \  |\  /\  ||           .-'".".
_.-'       '.   `'-._   .-'"/'.      "   | |/ /  | |/        _-"   (   '-_
             '.      _-"   (   '-_       \ | /   \ |     _.-'       )     "-._
           _.'   _.-'       )     "-._    ||\\   |\\  '"'        .-'
         '               .-'          `'  || \\  ||))
jjs__  _  ___  _ ____________ _____  ___ _|\ _|\_|\\/ _______________  ___   _
                       c  c  " c C ""C  " ""  "" ""
                   c       C
              C        C
                   C
    C     c


                 _____ _                 _           __            
                |_   _| |__   __ _ _ __ | | _____   / _| ___  _ __ 
                  | | | '_ \ / _` | '_ \| |/ / __| | |_ / _ \| '__|
                  | | | | | | (_| | | | |   <\__ \ |  _| (_) | |   
                  |_| |_| |_|\__,_|_| |_|_|\_\___/ |_|  \___/|_|   
                __      ____ _| |_ ___| |__ (_)_ __   __ _         
                \ \ /\ / / _` | __/ __| '_ \| | '_ \ / _` |        
                 \ V  V / (_| | || (__| | | | | | | | (_| |        
                  \_/\_/ \__,_|\__\___|_| |_|_|_| |_|\__, |        
                                                     |___/         


)")
